contracts/requirements.txt
python-dotenv
requests
eth-account
